Distribute test_posteriors into step_methods
